api/requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==0.5.0
Django==3.2.5
django-autoslug==1.9.8
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-timezone-field==4.2.1
djangorestframework==3.12.4
gunicorn==20.1.0
kombu==5.1.0
Pillow==7.0.0
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
